fix(F-06): move RepositoryProvider and RepositorySelectionOptions to Infrastructure
These types named concrete storage technologies (LocalJson, GoogleDriveJson) and carried
Google Drive-specific fields — pure Infrastructure concerns that had no business being in
the Application layer. IRepositoryFactory is also removed from Application: it was only
ever used inside InfrastructureServiceCollectionExtensions and added no value as an
Application abstraction.

Changes:
- RepositoryProvider and RepositorySelectionOptions moved to Financial.Infrastructure/Repositories
- IRepositoryFactory deleted from Application; RepositoryFactory now used directly
- InfrastructureServiceCollectionExtensions: removes IRepositoryFactory from DI, calls
  new RepositoryFactory().Create(options) inline; IServiceProvider parameter no longer needed
- RepositoryFactory: drops IRepositoryFactory implementation (interface deleted)
- RepositoryFactoryTests: updated using to Infrastructure namespace
