--- Financial.Application/Interfaces/IRepositoryFactory.cs
@@ -1,6 +1,0 @@
-namespace Financial.Application.Interfaces;
-
-public interface IRepositoryFactory
-{
-    IRepository Create(RepositorySelectionOptions options);
-}
